Use uint8_t to calculate poll delta

+++ arduino_code/src/main.cpp
@@ -113,10 +113,10 @@ static void fixButtonTiming(const uint16_t pressedAt) {
 
 // Controller will be polled twice per frame.
 const static uint8_t pollDelta = FRAME_LENGTH / 2;
-static uint16_t lastPoll;
+static uint8_t lastPoll;
 
 static void pollController() {
-    const uint16_t timeNow = millis();
+    const uint8_t timeNow = millis();
 
     if (timeNow - lastPoll > pollDelta) {
        ConsoleWriter.addData(ControllerReader.getData());
